fix 500 when delete orgnization and resolved #486

+++ models/user.go
@@ -891,8 +891,10 @@ func deleteUser(e *xorm.Session, u *User) error {
 	}
 
 	avatarPath := u.CustomAvatarPath()
-	if err := os.Remove(avatarPath); err != nil {
-		return fmt.Errorf("Fail to remove %s: %v", avatarPath, err)
+	if com.IsExist(avatarPath) {
+		if err := os.Remove(avatarPath); err != nil {
+			return fmt.Errorf("Fail to remove %s: %v", avatarPath, err)
+		}
 	}
 
 	return nil
